+++ src/prme/models/__init__.py
@@ -9,7 +9,7 @@ from prme.models.events import Event
 from prme.models.extraction import ExtractionRecord
 from prme.models.extraction_work import ExtractionStatus, ExtractionProcessingResult
 from prme.models.nodes import MemoryNode
-from prme.models.processing import ProcessingResult, ProcessingStatus
+from prme.models.processing import ProcessingResult, ProcessingStatus, StoreReceipt
 from prme.models.provenance import NodeProvenance, OperationAuditRecord
 from prme.models.profile import ProfileJobStatus, ProfileProcessingResult, ProfileCollectionResult
 
@@ -25,6 +25,7 @@ __all__ = [
     "OperationAuditRecord",
     "ProcessingResult",
     "ProcessingStatus",
+    "StoreReceipt",
     "ProfileJobStatus",
     "ProfileProcessingResult",
     "ProfileCollectionResult",